Wrote a query to create the table responded_applicants

+++ src/main/java/org/example/jobsearch/models/RespondApplicant.java
@@ -11,7 +11,7 @@ import lombok.*;
 @AllArgsConstructor
 public class RespondApplicant {
     private Long id;
-    private Integer resumeId;
-    private Integer vacancyId;
+    private Long resumeId;
+    private Long vacancyId;
     private Boolean confirmation;
 }
